Fixed snake growth
I was able to get snake growth working, found a few glitches though. The issue with disappearing on the bottom occurred on the right side as well, pellets will spawn on the snake, and if the snake loops on itself weird things will happen. That last one shouldn't be an issue however if we have snake death implemented