Fixed pellet consumption
Fixed pellet consumption so that pellets are actually being deleted and the snake will actually grow. Currently there is an issue however when visually the snake goes through the pellet but the pellet is not consumed. This is because the snake's x or y position is a multiple only 5 instead of 10. We may have to consider refactoring to get this to work. Additionally there are many graphical errors occurring.